+++ frontend/bindings/github.com/wenit/NetstatUI/services/monitor/index.ts
@@ -8,6 +8,7 @@ export {
 
 export {
     Diff,
+    GeoStatus,
     SnapshotResult,
     Stats
 } from "./models.js";